Added fantasy point calculation

+++ NHLweb/fantasy/models.py
@@ -70,7 +70,7 @@ class Stats(models.Model):
     
     evenTOI = models.CharField(max_length = 12, default = "N/A")
     
-    fantasyPoints = models.IntegerField(default = 0)
+    fantasyPoints = models.FloatField(default = 0)
     
 
     def __str__(self):
